client master

+++ api/app/Http/Controllers/Admin/ClientController.php
@@ -51,23 +51,23 @@ class ClientController extends Controller
         	return response()->json(['success' => false, 'errors' => $validator->errors()], 422);
         }
 
-        $client = Client::where('dni', $request['dni'])->where('dni', '!=', '900219765-2')->first();
+        $client = Client::where('dni', $request['dni'])->first();
 
         $doctor = Doctor::where('id', $request['doctor_id'])->first();
 
-        $clienteSession = [
-            ['client_id', '=', $client->id],
-            ['date', '=', $request['date']],
-        ];
-
-        $session = Session::where($clienteSession)->first();
-        
-        if ($session) {
-
-            $session->delete();
-        }
-        
         if ($client && $doctor) {
+
+            $clienteSession = [
+                ['client_id', '=', $client->id],
+                ['date', '=', $request['date']],
+            ];
+    
+            $session = Session::where($clienteSession)->first();
+            
+            if ($session && $client->dni != '900219765-2') {
+    
+                $session->delete();
+            }
 
             $attrSession = [
                 ['date', '=', $request['date']],
